Split out common shell, cmdline and tempdir helpers into pytest-shutil.
