--- .travis.yml
@@ -1,9 +1,0 @@
-language: python
-python:
-  - 2.6
-  - 2.7
-install:
-  - python setup.py develop
-  - pip list
-script:
-  - python setup.py test --unit
