--- pkglib-util/MANIFEST.in
@@ -1,1 +1,0 @@
-recursive-include tests *.py *.cpp *.hpp *.c *.h *.pyx
