+++ pytest-shutil/README.rst
@@ -1,0 +1,9 @@
+pytest-shutil
+================
+
+This library is a goodie-bag of Unix shell and environment management tools for automated tests. 
+                      
+Documentation
+-------------
+
+TODO
