+++ pytest-shutil/setup.cfg
@@ -7,3 +7,12 @@ tag_build = .dev1
 [upload_docs]
 upload-dir = build/sphinx/html
 show-response = 1
+
+[pytest]
+# This section sets configuration for all invocations of py.test,
+# both standalone cmdline and running via setup.py
+norecursedirs = 
+	.git 
+	*.egg 
+	build 
+	dist
