+++ pytest-shutil/setup.py
@@ -1,11 +1,13 @@
-import os, sys
+import sys
+import os
 from setuptools import setup, Command, find_packages
 
 
 classifiers = [
     'License :: OSI Approved :: MIT License',
-    'Development Status :: 4 - Beta',
+    'Development Status :: 5 - Production/Stable',
     'Topic :: Software Development :: Libraries',
+    'Topic :: Software Development :: Testing',
     'Topic :: Utilities',
     'Intended Audience :: Developers',
     'Operating System :: POSIX',
@@ -14,9 +16,10 @@ classifiers = [
     'Programming Language :: Python :: 2.7',
 ]
 
-long_description = "PkgLib common utilities"
+long_description = open("README.rst").read()
 
 trailing_args = []
+
 
 class Test(Command):
     user_options = [('unit', None, 'Just run unit tests'),
@@ -57,21 +60,24 @@ def main():
                 sys.argv = sys.argv[:-len(trailing_args)]
                 break
 
+    install_requires = ['pytest',
+                        'six',
+                        'execnet',
+                        ]
     setup(
-        name='pkglib-util',
-        description='PkgLib Utility Library',
+        name='pytest-shutil',
+        description='A goodie-bag of unix shell and environment tools for py.test',
         long_description=long_description,
-        version='0.10.6',
-        # url='',
+        version='1.0.0',
+        url='https://github.com/manahl/pytest-plugins',
         license='MIT license',
         platforms=['unix', 'linux'],
         author='Edward Easton',
         author_email='eeaston@gmail.com',
         classifiers=classifiers,
-        install_requires=[],
-        tests_require=['pytest'],
-        packages=find_packages(),
+        install_requires=install_requires,
         cmdclass={'test': Test},
+        packages=find_packages(),
     )
 
 if __name__ == '__main__':
