--- setup.cfg
@@ -1,8 +1,0 @@
-
-[multipkg]
-# These are the sub-package directories for this repository.
-# Setup.py commands will be run on each of these in order
-pkg_dirs =
-    pkglib-util
-    pkglib-testing
-    pkglib
