--- setup.py
@@ -1,6 +1,0 @@
-import sys
-sys.path.append('pkglib')
-sys.path.append('pkglib-testing')
-sys.path.append('pkglib-util')
-import pkglib.multipkg
-pkglib.multipkg.setup()
